requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.10
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
